Add bulk add support to addToAddressBook RPC method
